+++ wisplib/upper.c
@@ -1,8 +1,10 @@
 /* UPPER ... convert a string from lower case to upper case									*/
 
+#include "idsistd.h"
+
 UPPER(line,len)
 char *line;								/* the line of text to convert				*/
-long  *len;
+int4  *len;
 {
 	int i;
